requirements.txt
bleach==6.3.0
bokeh==3.8.1
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
colorcet==3.1.0
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
holoviews==1.22.1
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
linkify-it-py==2.0.3
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdit-py-plugins==0.5.0
mdurl==0.1.2
narwhals==2.14.0
networkx==3.6.1
numpy==2.4.0
packaging==25.0
pandas==2.3.3
panel==1.8.5
param==2.3.1
pillow==12.0.0
plotly==6.5.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-louvain==0.16
pytz==2025.2
pyviz_comms==3.0.6
PyYAML==6.0.3
rdflib==7.5.0
redis==7.1.0
requests==2.32.5
seaborn==0.13.2
six==1.17.0
SPARQLWrapper==2.0.0
spotipy==2.25.2
tornado==6.5.4
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==1.0.3
urllib3==2.6.2
webencodings==0.5.1
xyzservices==2025.11.0
